backend/requirements.txt
# Django Core - Latest LTS with full compatibility
Django==5.2.5
djangorestframework==3.16.1
django-cors-headers==4.6.0
django-environ==0.11.2

# Database
psycopg2-binary==2.9.9
dj-database-url==2.1.0

# Authentication & Security
djangorestframework-simplejwt==5.3.0
django-allauth==0.63.6
django-ipware==7.0.1

# API Documentation
drf-yasg==1.21.7

# Background Jobs - Latest stable versions
celery==5.5.3
redis>=5.2.1
django-celery-beat==2.8.1

# File Storage
boto3==1.35.19
django-storages==1.14.4
Pillow==10.4.0

# Integrations - Latest versions
stripe==12.4.0
requests==2.32.3

# Monitoring & Logging
sentry-sdk==2.13.0

# Development
django-debug-toolbar==4.4.6
django-extensions==3.2.3

# Testing (Updated October 2025)
pytest==8.3.3
pytest-django==4.9.0
pytest-cov==5.0.0
factory-boy==3.3.1
faker==30.8.2
freezegun==1.5.1

# Production Server
gunicorn==23.0.0
whitenoise==6.7.0

# Rate Limiting & Caching
django-ratelimit==4.1.0
django-redis==5.4.0

# AI Assistant
langchain-anthropic>=0.3.0
langchain-core>=0.3.0
langgraph>=0.4.0
langgraph-checkpoint-redis>=0.3.0
langsmith>=0.3.0
gevent>=24.0